+++ framework/domain/src/main/java/is/codion/framework/domain/entity/attribute/Column.java
@@ -19,6 +19,7 @@
 package is.codion.framework.domain.entity.attribute;
 
 import is.codion.common.db.database.Database;
+import is.codion.common.utilities.TypeReference;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.EntityType;
 import is.codion.framework.domain.entity.attribute.DefaultColumn.DefaultColumnDefiner;